+++ Source/Core/PixelFormat.js
@@ -23,7 +23,7 @@ define([
         DEPTH_COMPONENT : WebGLConstants.DEPTH_COMPONENT,
 
         /**
-         * A pixel format containing a depth and stencil value, most often used with {@link PixelDatatype.UNSIGNED_INT_24_8_WEBGL}.
+         * A pixel format containing a depth and stencil value, most often used with {@link PixelDatatype.UNSIGNED_INT_24_8}.
          *
          * @type {Number}
          * @constant