+++ Source/Scene/BillboardCollection.js
@@ -552,7 +552,9 @@ define([
             return indexBuffer;
         }
 
-        var length = sixteenK * 6;
+        // Subtract 6 because the last index is reserverd for primitive restart.
+        // https://www.khronos.org/registry/webgl/specs/latest/2.0/#5.18
+        var length = sixteenK * 6 - 6;
         var indices = new Uint16Array(length);
         for (var i = 0, j = 0; i < length; i += 6, j += 4) {
             indices[i] = j;
@@ -1133,7 +1135,7 @@ define([
         }
     }
 
-    var scratchDrawingBufferDimensions = new Cartesian2();
+    var scratchPixelSize = new Cartesian2();
     var scratchToCenter = new Cartesian3();
     var scratchProj = new Cartesian3();
     function updateBoundingVolume(collection, context, frameState, boundingVolume) {
@@ -1146,9 +1148,7 @@ define([
             var proj = Cartesian3.multiplyByScalar(camera.directionWC, Cartesian3.dot(toCenter, camera.directionWC), scratchProj);
             var distance = Math.max(0.0, Cartesian3.magnitude(proj) - boundingVolume.radius);
 
-            scratchDrawingBufferDimensions.x = context.drawingBufferWidth;
-            scratchDrawingBufferDimensions.y = context.drawingBufferHeight;
-            var pixelSize = frustum.getPixelSize(scratchDrawingBufferDimensions, distance);
+            var pixelSize = frustum.getPixelDimensions(context.drawingBufferWidth, context.drawingBufferHeight, distance, scratchPixelSize);
             pixelScale = Math.max(pixelSize.x, pixelSize.y);
         }